requirements.txt
google-generativeai>=0.3.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
duckduckgo-search>=3.9.0
pydantic>=1.8.2
python-multipart>=0.0.5
requests>=2.31.0 